Move `SvgColorControl` to svg-icon components
