+++ src/blocks/svg-icon-maxi/components/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as SvgColorControl } from './svg-color-control';
