+++ src/components/index.js
@@ -64,6 +64,7 @@ export { default as SettingTabsControl } from './setting-tabs-control';
 export { default as ShapeDivider } from './shape-divider';
 export { default as Spinner } from './spinner';
 export { default as SvgAltControl } from './svg-alt-control';
+export { default as SvgColor } from './svg-color';
 export { default as SVGFillControl } from './svg-fill-control';
 export { default as SvgStrokeWidthControl } from './svg-stroke-width-control';
 export { default as SvgWidthControl } from './svg-width-control';
@@ -76,4 +77,3 @@ export { default as TransformControl } from './transform-control';
 export { default as TransitionControl } from './transition-control';
 export { default as TypographyControl } from './typography-control';
 export { default as ZIndexControl } from './zindex-control';
-export * from './svg-color';
